Reimplementation of the core estimation algorithm using constraints from Tabesh et. al. 2011.
Improved results using the constrained estimation approach developed in
Tabesha, Jensena, Ardekanic, Helpern
Estimation of Tensors and Tensor-Derived Measures in Diffusional Kurtosis Imaging
Magn Reson Med. 2011 March ; 65(3): 823–836. doi:10.1002/mrm.22655.

The only difference is an interior-point method is used, rather than
an active set method, to solve the quadratic programming problem
given by the constrained least squares solution to the tensor fitting.

This improves the results, however the implementation is still rather
slow and optimizations are needed. The computation of the scalar
diffusion kurtosis is also still the simple form which generates
singularities. The above paper also presents an alternative that
needs to be investigated.