Detect nonfinite matrices in interior point and halt early
In real data, particularly in the BG where SNR is low, the Newton
iterations inside the interior point method become unstable. This
is now detected and the estimation halts early.

A side effect is that the initial condition can no longer be the
last solution, since it might be bad, so the initial condition is
now the same for all voxels.